+++ gpt_lib/nexus.py
@@ -1,0 +1,68 @@
+import os
+import gpt_lib.tools as tools
+from gpt_lib.openai_wrapper import gpt3_embedding, gpt3_completion
+from time import time
+from uuid import uuid4
+
+#
+# Nexus methods for conversation
+# 
+#
+
+def load_convo():
+    files = os.listdir('nexus')
+    files = [i for i in files if '.json' in i]  # filter out any non-JSON files
+    result = list()
+    for file in files:
+        data = tools.load_json('nexus/%s' % file)
+        result.append(data)
+    ordered = sorted(result, key=lambda d: d['time'], reverse=False)  # sort them all chronologically
+    return ordered
+
+def save_conversation(speaker, msg, conversations):
+    timestamp = time()
+    vector = gpt3_embedding(msg)
+    timestring = tools.timestamp_to_datetime(timestamp)
+    message = '%s: %s - %s' % (speaker, timestring, msg)
+
+    info = {'speaker': speaker, 'time': timestamp, 'vector': vector, 'message': message, 'uuid': str(uuid4()), 'timestring': timestring}
+    filename = 'log_%s_%s.json' % (timestamp, speaker)
+    tools.save_json('nexus/%s' % filename, info)
+    conversations.append(info)
+
+    return vector
+
+def get_last_messages(conversation, limit):
+    try:
+        short = conversation[-limit:]
+    except:
+        short = conversation
+    output = ''
+    for i in short:
+        output += '%s\n\n' % i['message']
+    output = output.strip()
+    return output
+
+def fetch_memories(vector, logs, count):
+    scores = list()
+    for i in logs:
+        if vector == i['vector']:
+            # skip this one because it is the same message
+            continue
+        score = tools.similarity(i['vector'], vector)
+        i['score'] = score
+        scores.append(i)
+    ordered = sorted(scores, key=lambda d: d['score'], reverse=True)
+    # TODO - pick more memories temporally nearby the top most relevant memories
+    try:
+        ordered = ordered[0:count]
+        return ordered
+    except:
+        return ordered
+
+def ask_theo(note, recent):
+    prompt = tools.open_file('promts/prompt_response.txt').replace('<<NOTES>>', note).replace('<<CONVERSATION>>', recent)
+    #### generate response, vectorize, save, etc
+    output = gpt3_completion(prompt)
+
+    return output
