+++ gpt_lib/openai_wrapper.py
@@ -1,0 +1,41 @@
+import openai
+import re
+import os
+from gpt_lib import tools
+from time import sleep, time
+
+def gpt3_completion(prompt, engine='text-davinci-003', temp=0.0, top_p=1.0, tokens=400, freq_pen=0.0, pres_pen=0.0, stop=['USER:', 'THEO:']):
+    max_retry = 5
+    retry = 0
+    prompt = prompt.encode(encoding='ASCII',errors='ignore').decode()
+    while True:
+        try:
+            response = openai.Completion.create(
+                engine=engine,
+                prompt=prompt,
+                temperature=temp,
+                max_tokens=tokens,
+                top_p=top_p,
+                frequency_penalty=freq_pen,
+                presence_penalty=pres_pen,
+                stop=stop)
+            text = response['choices'][0]['text'].strip()
+            text = re.sub('[\r\n]+', '\n', text)
+            text = re.sub('[\t ]+', ' ', text)
+            filename = '%s_gpt3.txt' % time()
+            if not os.path.exists('gpt3_logs'):
+                os.makedirs('gpt3_logs')
+            tools.save_file('gpt3_logs/%s' % filename, prompt + '\n\n==========\n\n' + text)
+            return text
+        except Exception as oops:
+            retry += 1
+            if retry >= max_retry:
+                return "GPT3 error: %s" % oops
+            print('Error communicating with OpenAI:', oops)
+            sleep(1)
+
+def gpt3_embedding(content, engine='text-embedding-ada-002'):
+    content = content.encode(encoding='ASCII',errors='ignore').decode()
+    response = openai.Embedding.create(input=content,engine=engine)
+    vector = response['data'][0]['embedding']  # this is a normal list
+    return vector
