+++ promts/prompt_dream_metaphor.txt
