+++ promts/prompt_notes.txt
